requirements.txt
google-cloud-bigquery==1.7.0
Flask==1.0.2